BUG#17233 : LOAD DATA INFILE: failure causes mysqld dbug_assert, binlog not flushed
This patch corrects a bug involving a LOAD DATA INFILE operation on a 
transactional table. It corrects a problem in the error handler moving
the transactional table check and autocommit_or_rollback operation to the 
end of the error handler. An additional test case was added to detect this
condition.