WL#3570 : Default algorithm backup & restore (blocking)
This patch implements the default backup blocking algorithm. The patch
also includes enhancements the kernel for better open tables behavior 
for backup and better open and lock methods for restore.

The patch also contains the Buffer_iterator class and header files. The
Buffer_iterator class is used to buffer data during read and write that
is larger than the kernel's stream buffer.