Bug#29451 (Broken makefile dependencies with libndb.la)
In Automake,
mysqld_LDADD = libndb.la
only adds libndb to the link command for mysqld,
but does not declare a dependency.

Added libndb.la to mysqld_DEPENDENCIES

This fix a build race condition that currently
breaks make -J builds, and also enforce a re-link
of mysqld when libndb.la changes.