Fix a bad BitKeeper dependency structure for the "funcs_2" suite:
The files below "mysql-test/suite/funcs_2" in version 5.1
did not depend on the equivalent ones in 5.0,
probably because they had been checked in independent of each other
in both versions.

Foreach file F in the suite that has a "deleted" counterpart D, use
    bk rm $F
    bk mv $D $F
    bk edit $F
to get those files into the 5.1 suite that (for BK) depend on 5.0.

There is only one file whose contents differs between 5.0 and 5.1,
restore its current 5.1 contents:
   mysql-test/suite/funcs_2/r/ndb_charset.result