Bug#30793: set random seed for mysqlslap
Contributed patch from "mushuiliu".  No CLA.

Add a new parameter "set-random-seed", which takes an uint
as an argument.

If the parameter isn't used or is used for value "0", then the
random_seed is set from time() as usual.

Print the seed value in the output, so a run can be replayed.