Bug#28870 check that table locks are released/reset
The problem is that some mysql_lock_tables error paths are not resetting the
tables lock type back to TL_UNLOCK. If the lock types are not reset properly,
a table might be returned to the table cache with wrong lock_type.
    
The proposed fix is to ensure that the tables lock type is always properly
reset when mysql_lock_tables fails. This is a incompatible change with respect
to the process state information.