BUG #8368
"mysqldump needs --slave-data option"

Added this option, named as "--dump-slave". The purpose of this option is to be
able to produce a dump from a slave used for making backups of the master. Originally,
dumping from the main master was fine, but as more data accumulated, the
dump process would take over 30 minutes, locking up the master database hence website
for 30 minutes.  A slave dedicated to producing backups was the answer, but I needed a
dump that could be used to restore a slave instantly and in order to do that, it has to
have three things contained in the dump:

1. "STOP SLAVE;" at the beginning
2. "CHANGE MASTER TO ...<the master - info from 'show slave status'>"
3. "START SLAVE;" at the end

These options in this changeset contain this.

--apply-slave-statements adds "STOP SLAVE" to the beginning of the dump and "START SLAVE" 
to the end of the dump.

--include-master-host-port gives the user the option to have the host explicitely
added to the "CHANGE MASTER TO ..." line.

--dump-slave adds the "CHANGE MASTER ..." to the dump representing not
the slave's master binlog info, but the slave's master's info from "SHOW SLAVE STATUS"