Bug #30695: An apostrophe ' in the comment of the ADD PARTITION   causes the Server to crash.
Crashed because of the comment in the partition clause was saved in
the frm file without escaping, causing the server to crash when it was
read/parsed again.

Fixed by escaping quoted text in the partition info when writing it to
the frm-file and added a check that it was able to parse the partition
info before using it 

NOTE: If the comment is written by an earlier version of the server,
the corrupted frm-file is not fixed, but left corrupted, you have to
manually drop the table and recreate it.