Bug#31608 missing mysqltest change_user command
The problem is that currently there is no way to test the behavior
of the mysql_change_user() function using the mysqltest suite because
there is no internal command for it.

The solution is to introduce a change_user command that can be used
to test aspects of the MySQL client function mysql_change_user().