6.0 version of the fix for bug #25421: MySQL threads don't respond to the kill command.
Calculating the estimated number of records for a range scan may take a
significant time, and it was impossible for a user to interrupt that
process by killing the connection or the query.

Fixed by checking the thread's 'killed' status in 
handler::multi_range_read_info_const() and interrupting the calculation 
process if it is set to a non-zero value. The same must also be done in
every derived class that overrides multi_range_read_info_const().