Bug#32078 - Excessive warnings: One can only use the --user switch             if running as root
Every start of a server in the test suite raised that warning.

The cause was an unconditionla add of the --user option to the
server command line. Only the "root" user (effective user id == 0)
must use that option.

Added check for effective user id == 0 before adding --user.

Thanks to Magnus Svensson for the patch.