BUG#31383 : Consistent Snapshot driver not consistent
This patch changes the default driver to use a separate thread to open and lock tables.
The default driver opens tables on the prelock() call from the kernel. The snapshot
driver initiates the CS read on the lock() call from the kernel and opens tables in the first
call to get_data() after the lock is taken. This is due to the fact that the default driver's
validity point is at open_and_lock_tables() while the snapshot driver's validity point
is at the start of the transaction.