WL#4062 : Online Backup: Metadata freeze
This patch implements the first iteration of the metadata freeze (hence
DDL blocker) for the online backup project. It uses a normal pthread
condition variable loop to block all DDL operations while a backup is
running and another to block a backup if any DDL operations are running.

Note: This patch requires the patch for WL#4120.