Fix mysql-6.0-rpl-merge pushbuild failures
Tests that use get_lock(), BACKUP_BREAKPOINT(), and release_lock() do
not run correctly under valgrind. The BACKUP_BREAKPOINT() macro is not
defined in the build so the get_lock() fails to pause the code at the
breakpoint. This causes timing issues with wait_condition.inc timing
loops in the test which leads to test failures on mismatched results.
Thus, these tests cannot be run on valgrind.