BUG#33119 (Online backup silently overwrites files).
This patch makes backup to fail with error "Can't write to backup location 
'<path>'" if the file already exists. Backup tests are updated so that they 
don't try to overwrite existing files.

Also, in case of failed BACKUP operation, the backup file is deleted if it was 
created at its beginning.