Bug#27427: resolveip fails on hostnames with a leading digit
Patch by Kasper Dupont.  No CLA required for this size of patch.

"resolveip" program produces incorrect result if given a hostname
starting with a digit.  Someone seems to have thought that names 
can not have digits at the beginning.

Instead, use the resolver library to work out the rules of hostnames, 
as it will undoubtedly be better at it than we are.